test: increase tsc-compliance test coverage (#12444)
