+++ scripts/tsc-compliance-test/tsconfig.ts4_2.json
@@ -9,5 +9,5 @@
 		"moduleResolution": "node",
 		"types": ["node"]
 	},
-	"include": ["ts4_2.ts"]
+	"include": ["publicPaths.ts"]
 }
